Moved bot files to bot directory for organization
