+++ bot/cogs.py
